Splits bundle into mutliple entries to support tree shaking
