+++ src/setupWorker/glossary.ts
